+++ provider/registry/src/lib.rs
@@ -199,8 +199,8 @@ macro_rules! registry(
             icu::segmenter::provider::SentenceBreakDataV2: SentenceBreakDataV2,
             icu::segmenter::provider::WordBreakDataOverrideV1: WordBreakDataOverrideV1,
             icu::segmenter::provider::WordBreakDataV2: WordBreakDataV2,
-            icu::time::provider::names::Bcp47ToIanaMapV1: Bcp47ToIanaMapV1,
-            icu::time::provider::names::IanaToBcp47MapV3: IanaToBcp47MapV3,
+            icu::time::provider::iana::Bcp47ToIanaMapV1: Bcp47ToIanaMapV1,
+            icu::time::provider::iana::IanaToBcp47MapV3: IanaToBcp47MapV3,
             icu::time::provider::windows::WindowsZonesToBcp47MapV1: WindowsZonesToBcp47MapV1,
             icu::datetime::provider::neo::WeekdayNamesV1: WeekdayNamesV1,
             icu::datetime::provider::neo::DayPeriodNamesV1: DayPeriodNamesV1,
